misc: addressed comments
